Added CHECKSUM_G1_1_08k_4GB to version check
TLDR; This change is required to run the Gothic Reloaded Mod.

D3D11 failed to allocate memory for a vertex buffer (D3D11ConstantBuffer.cpp, line 24 returned E_OUTOFMEMORY) and then the renderer crashed to a black screen. The Gothic process was around 600-700 MB RAM, far from the 2GB limit, while my system has 32 GB of CPU memory and 8 GB of GPU memory, both being about 80% free. My assumption was that the vertex buffer to be created, which is around 800 MB for some reason, is to be allocated in the process memory because D3D11_USAGE_DYNAMIC is specified and this fails as the memory is fragmented and there is not enough _sequential_ free memory. As a workaround I used the large address aware hack from NicoDE, which seems to have solved the problem. Unfortunately this caused the renderer to complain about the checksum.